pyproject.toml
[project]
name = "production-research-rag"
version = "0.1.0"
description = "A production-ready RAG pipeline for research papers"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "ipykernel>=7.2.0",
    "langchain>=1.2.10",
    "langchain-community>=0.4.1",
    "langchain-core>=1.2.17",
    "langchain-openai>=0.3.0",
    "langchain-unstructured>=1.0.1",
    "ml-collections>=1.1.0",
    "openai>=2.24.0",
    "pandas>=2.3.3",
    "pdfminer>=20191125",
    "pinecone>=8.1.0",
    "pinecone-text>=0.11.0",
    "prefect>=3.6.20",
    "pydantic>=2.12.5",
    "pypdfium>=0.0.15",
    "python-dotenv>=1.2.1",
    "pyyaml>=6.0.3",
    "streamlit>=1.54.0",
    "tiktoken>=0.12.0",
    "unstructured-client>=0.42.10",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
]
